Refactor version handling into separate class and add support for prerelease SDK
