+++ aemanalyser-maven-plugin/src/test/java/com/adobe/aem/analyser/mojos/AemAnalyseMojoTest.java
@@ -20,17 +20,11 @@ import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.model.Build;
-import org.apache.maven.model.Dependency;
-import org.apache.maven.model.DependencyManagement;
-import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.project.MavenProject;
 import org.apache.sling.feature.ArtifactId;
 import org.apache.sling.feature.builder.ArtifactProvider;
@@ -104,79 +98,9 @@ public class AemAnalyseMojoTest {
         assertEquals(def456.toUri().toURL(), url2);
     }
 
-    @Test
-    public void testGetArtifactIdFromDependencies2() throws Exception {
-        Dependency dep = new Dependency();
-        Dependency sdkDep = new Dependency();
-        sdkDep.setGroupId("com.adobe.aem");
-        sdkDep.setArtifactId("aem-sdk-api");
-        sdkDep.setVersion("1.3.5");
-
-        DependencyManagement depMgmt = new DependencyManagement();
-        depMgmt.addDependency(dep);
-        depMgmt.addDependency(sdkDep);
-
-        MavenProject prj = Mockito.mock(MavenProject.class);
-        Mockito.when(prj.getDependencyManagement()).thenReturn(depMgmt);
-
-        AemAnalyseMojo mojo = new TestAnalyseMojo(prj);
-
-        assertEquals(new ArtifactId(sdkDep.getGroupId(), sdkDep.getArtifactId(), sdkDep.getVersion(), sdkDep.getClassifier(), sdkDep.getType()), 
-                mojo.getArtifactIdFromDependencies("com.adobe.aem", "aem-sdk-api"));
-    }
-
-    @Test
-    public void testDiscoverAddons() throws Exception {
-        Dependency dep = new Dependency();
-        Dependency sdkDep = new Dependency();
-        sdkDep.setGroupId("com.adobe.aem");
-        sdkDep.setArtifactId("aem-cif-sdk-api");
-        sdkDep.setClassifier("aem-cif-sdk");
-        sdkDep.setVersion("1.3.2");
-
-        DependencyManagement depMgmt = new DependencyManagement();
-        depMgmt.addDependency(dep);
-        depMgmt.addDependency(sdkDep);
-
-        MavenProject prj = Mockito.mock(MavenProject.class);
-        Mockito.when(prj.getDependencyManagement()).thenReturn(depMgmt);
-
-        AemAnalyseMojo mojo = new TestAnalyseMojo(prj);
-        final List<ArtifactId> addons = mojo.discoverAddons(null);
-        assertEquals(1, addons.size());
-        assertEquals(new ArtifactId(sdkDep.getGroupId(), sdkDep.getArtifactId(), sdkDep.getVersion(), sdkDep.getClassifier(), sdkDep.getType()), 
-                addons.get(0));
-    }
-    
-    @Test
-    public void testGetSDKArtifactIdWithConfiguredVersionAndArtifactId() throws Exception {
-        final MavenProject prj = Mockito.mock(MavenProject.class);
-        final AemAnalyseMojo mojo = new TestAnalyseMojo(prj);
-        mojo.sdkArtifactId = "my-artifact";
-        mojo.sdkVersion = "5.0";
-
-        final ArtifactId id = mojo.getSDKArtifactId();
-        assertEquals(Constants.SDK_GROUP_ID, id.getGroupId());
-        assertEquals("my-artifact", id.getArtifactId());
-        assertEquals("5.0", id.getVersion());
-    }
-
-    @Test(expected = MojoExecutionException.class)
-    public void testGetSDKArtifactIdWithoutDependency() throws Exception {
-        final MavenProject prj = Mockito.mock(MavenProject.class);
-        final AemAnalyseMojo mojo = new TestAnalyseMojo(prj);
-        mojo.sdkArtifactId = Constants.SDK_ARTIFACT_ID;
-
-        mojo.getSDKArtifactId();
-    }
-
     private static class TestAnalyseMojo extends AemAnalyseMojo {
         private TestAnalyseMojo(final MavenProject prj) {
             this.project = prj;
-        }
-
-        String getLatestVersion(final Dependency dependencies) throws MojoExecutionException {
-            return null;
         }
     }
 }
